iNethack2 update

+++ nethack/src/region.c
@@ -1,4 +1,4 @@
-/*	SCCS Id: @(#)region.c	3.4	2002/10/15	*/
+/*	SCCS Id: @(#)region.c	5.4	2002/10/15	*/
 /* Copyright (c) 1996 by Jean-Christophe Collet	 */
 /* NetHack may be freely redistributed.  See license for details. */
 
@@ -878,10 +878,10 @@ genericptr_t p1;
 genericptr_t p2;
 {
     NhRegion *reg;
-    int damage;
+    size_t damage;
 
     reg = (NhRegion *) p1;
-    damage = (int) reg->arg;
+    damage = (size_t) reg->arg;
 
     /* If it was a thick cloud, it dissipates a little first */
     if (damage >= 5) {
@@ -903,7 +903,7 @@ genericptr_t p2;
     int dam;
 
     reg = (NhRegion *) p1;
-    dam = (int) reg->arg;
+    dam = (size_t) reg->arg;
     if (p2 == NULL) {		/* This means *YOU* Bozo! */
 	if (nonliving(youmonst.data) || Breathless)
 	    return FALSE;
@@ -951,7 +951,7 @@ NhRegion *
 create_gas_cloud(x, y, radius, damage)
 xchar x, y;
 int radius;
-int damage;
+size_t damage;
 {
     NhRegion *cloud;
     int i, nrect;
